requirements.txt
BeautifulSoup==3.2.0
Django==1.4.1
Pillow==1.7.7
South==0.8.1
distribute==0.7.3
django-appmedia==1.0.1
django-autoslug==1.6.1
django-ckeditor==3.6.2.1
django-classy-tags==0.4
django-cms==2.3.3
django-debug-toolbar==0.9.4
django-grappelli==2.4.0
django-mptt==0.5.2
django-sekizai==0.7
html5lib==1.0b2
mailsnake==1.6.1
requests==1.2.3
simplejson==3.3.0
six==1.3.0
sorl-thumbnail==11.12
wsgiref==0.1.2
